is_idle() をインライン化
