+++ .projectile
@@ -1,2 +1,4 @@
+-backup
 -build
+-intermediate
 -profiler.log
