+++ test/block_fall_utest.lua
@@ -55,7 +55,7 @@ describe('ブロックの落下', function()
       board:update()
       board:update()
 
-      assert.is_true(block:is_idle())
+      assert.is_true(block._state == "idle")
     end)
   end)
 
@@ -118,8 +118,8 @@ describe('ブロックの落下', function()
       board:update()
       board:update()
 
-      assert.is_true(block1:is_idle())
-      assert.is_true(block2:is_idle())
+      assert.is_true(block1._state == "idle")
+      assert.is_true(block2._state == "idle")
     end)
   end)
 
@@ -179,8 +179,8 @@ describe('ブロックの落下', function()
       board:update()
       board:update()
 
-      assert.is_true(control:is_idle())
-      assert.is_true(cnot_x:is_idle())
+      assert.is_true(control._state == "idle")
+      assert.is_true(cnot_x._state == "idle")
     end)
   end)
 
@@ -239,8 +239,8 @@ describe('ブロックの落下', function()
       board:update()
       board:update()
 
-      assert.is_true(swap_left:is_idle())
-      assert.is_true(swap_right:is_idle())
+      assert.is_true(swap_left._state == "idle")
+      assert.is_true(swap_right._state == "idle")
     end)
 
     it('下のブロックをずらして落とす', function()
@@ -270,8 +270,8 @@ describe('ブロックの落下', function()
 
       assert.are_equal(swap_left, board:block_at(1, 1))
       assert.are_equal(swap_right, board:block_at(2, 1))
-      assert.is_true(swap_left:is_idle())
-      assert.is_true(swap_right:is_idle())
+      assert.is_true(swap_left._state == "idle")
+      assert.is_true(swap_right._state == "idle")
     end)
   end)
 
@@ -323,7 +323,7 @@ describe('ブロックの落下', function()
       board:update()
       board:update()
 
-      assert.is_true(garbage:is_idle())
+      assert.is_true(garbage._state == "idle")
     end)
   end)
 end)
