+++ test/i_block_utest.lua
@@ -8,9 +8,9 @@ describe('i_block', function()
     i = block_class("i")
   end)
 
-  describe("is_idle", function()
+  describe("_state", function()
     it("should return true", function()
-      assert.is_true(i:is_idle())
+      assert.is_true(i._state == "idle")
     end)
   end)
