+++ test/reduction_rules_utest.lua
@@ -1,6 +1,5 @@
 require("engine/test/bustedhelper")
 require("test/test_helper")
-require("lib/cursor")
 require("lib/block")
 require("lib/board")
