+++ test/x_block_utest.lua
@@ -20,9 +20,9 @@ describe('x_block', function()
     end)
   end)
 
-  describe(":is_idle", function()
+  describe("_state", function()
     it("should return true", function()
-      assert.is_true(x:is_idle())
+      assert.is_true(x._state == "idle")
     end)
   end)
